m:排查到bufferdata的binding函数实现有误导致三角形绘制不正常

+++ node/binding/webGLRes/WebGLBuffer.cc
@@ -13,7 +13,7 @@ namespace NodeBinding
 Napi::FunctionReference WebGLBuffer::constructor;
 WebGLBuffer::WebGLBuffer(const Napi::CallbackInfo &info) : Napi::ObjectWrap<WebGLBuffer>(info)
 {
-    this->mId = info[0].As<Napi::Number>().FloatValue();
+    this->mId = info[0].As<Napi::Number>().Uint32Value();
 }
 
 void WebGLBuffer::Init(Napi::Env env)
